Register

+++ ApiFinalPr/Apps/AdminApi/Controllers/BookController.cs
@@ -2,10 +2,12 @@
 using ApiFinalPr.Apps.AdminApi.DTOs.BookDtos;
 using ApiFinalPr.Data.DAL;
 using ApiFinalPr.Data.Entities;
+using AutoMapper;
 using EduHomeBEProject.Extensions;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -19,18 +21,20 @@ namespace ApiFinalProject.Data.DAL.Entities
     {
         private readonly AppDbContext _context;
         private readonly IWebHostEnvironment _env;
+        private readonly IMapper _mapper;
 
-        public BooksController(AppDbContext context, IWebHostEnvironment env)
+        public BooksController(AppDbContext context, IWebHostEnvironment env,IMapper mapper)
         {
             _context = context;
             _env = env;
+            _mapper = mapper;
         }
 
 
         [HttpGet("{id}")]
         public IActionResult Get(int id)
         {
-            Book book = _context.Books.FirstOrDefault(b => b.Id == id);
+            Book book = _context.Books.Include(b=>b.Author).ThenInclude(a=>a.Books).Include(b=>b.Genre).ThenInclude(g=>g.Books).FirstOrDefault(b => b.Id == id);
             if (book == null) return NotFound();
             BookGetDto bookDto = new BookGetDto
             {
@@ -41,7 +45,20 @@ namespace ApiFinalProject.Data.DAL.Entities
                 DisplayStatus = book.DisplayStatus,
                 Image = book.Image,
                 CreatedAt = book.CreatedAt,
-                ModifiedAt = book.ModifiedAt
+                ModifiedAt = book.ModifiedAt,
+                Author=new AuthorInProductGetDto
+                {
+                    AuthorId=book.Author.Id,
+                    AuthorName=book.Author.Name,
+                    BooksCount=book.Author.Books.Count
+                    
+                },
+                Genre=new GenreInProductGetDto
+                {
+                    GenreId = book.Genre.Id,
+                    GenreName = book.Genre.Name,
+                    BooksCount=book.Genre.Books.Count
+                }
             };
 
             return StatusCode(200, bookDto);
@@ -49,7 +66,7 @@ namespace ApiFinalProject.Data.DAL.Entities
         [HttpGet("")]
         public IActionResult GetAll(int page = 1, string search = null)
         {
-            var query = _context.Books.Where(x => !x.IsDeleted);
+            var query = _context.Books.Include(b=>b.Author).Include(b=>b.Genre).Where(x => !x.IsDeleted);
             if (!string.IsNullOrWhiteSpace(search))
                 query = query.Where(x => x.Name.Contains(search));
             ListDto<BookListItemDto> listDto = new ListDto<BookListItemDto>
@@ -63,7 +80,18 @@ namespace ApiFinalProject.Data.DAL.Entities
                     Cost = x.Cost,
                     Image = x.Image,
                     DisplayStatus = x.DisplayStatus,
-                    AuthorId=x.AuthorId
+                    Author=new AuthorInProductListItemDto
+                    {
+                        Id=x.Author.Id,
+                        Name=x.Author.Name
+                    },
+                    Genre=new GenreInProductListItemDto
+                    {
+                        Id=x.Genre.Id,
+                        Name=x.Genre.Name
+                    }
+                    //AuthorId=x.AuthorId,
+                    //GenreId=x.GenreId
                 }
                 ).ToList(),
                 TotalCount = query.Count()
@@ -74,13 +102,18 @@ namespace ApiFinalProject.Data.DAL.Entities
         [HttpPost("")]
         public IActionResult Create([FromForm] BookCreateDto bookDto)
         {
+
+            if (!_context.Authors.Any(a => a.Id == bookDto.AuthorId && !a.IsDeleted))
+                return NotFound();
+
             Book book = new Book
             {
                 Name = bookDto.Name,
                 Price = bookDto.Price,
                 Cost = bookDto.Cost,
                 DisplayStatus = bookDto.DisplayStatus,
-                AuthorId=bookDto.AuthorId
+                AuthorId=bookDto.AuthorId,
+                GenreId=bookDto.GenreId
             };
             book.Image = bookDto.ImageFile.SaveImg(_env.WebRootPath, "assets/img");
 
@@ -100,6 +133,9 @@ namespace ApiFinalProject.Data.DAL.Entities
             Book existBook = _context.Books.FirstOrDefault(b => b.Id == id);
             if (existBook == null) return NotFound();
 
+            if ((existBook.AuthorId!=bookDto.AuthorId && !_context.Authors.Any(a => a.Id == bookDto.AuthorId && !a.IsDeleted))|| 
+                (existBook.GenreId != bookDto.GenreId && !_context.Genres.Any(g => g.Id == bookDto.GenreId && !g.IsDeleted))) return NotFound();
+
             ApiFinalPr.Helpers.Helper.DeleteImg(_env.WebRootPath, "assets/img", existBook.Image);
             existBook.Image = bookDto.ImageFile.SaveImg(_env.WebRootPath, "assets/img");
             existBook.Name = bookDto.Name;
@@ -107,6 +143,7 @@ namespace ApiFinalProject.Data.DAL.Entities
             existBook.Cost = bookDto.Cost;
             existBook.DisplayStatus = bookDto.DisplayStatus;
             existBook.AuthorId = bookDto.AuthorId;
+            existBook.GenreId = bookDto.GenreId;
             //existBook.IsDeleted = bookDto.IsDeleted;
             _context.SaveChanges();
             return NoContent();
@@ -118,7 +155,7 @@ namespace ApiFinalProject.Data.DAL.Entities
             Book book = _context.Books.FirstOrDefault(b => b.Id == id);
             if (book == null) return NotFound();
 
-            _context.Books.Remove(book);
+            book.IsDeleted=true;
             _context.SaveChanges();
             return NoContent();
         }
